feat: Add task history records when collaborator is added or removed
This commit adds the functionality to create task history records when a collaborator is added or removed from a task. A new field has been added to the TaskHistory model called "Collaborator Added" and "Collaborator Deleted," which are now part of the enum of allowed actions for a TaskHistory object.